working events

+++ src/components/event/EventCard.js
@@ -9,22 +9,35 @@ class EventCard extends Component {
     //     return <EditModalHelper />
     // }
 
-  render() {
-    return (
-      <div className="card">
-        <div className="card-content">
-          <h3>{this.props.event.eventName}</h3>
-          <p>Location: {this.props.event.location}</p>
-          <p>Date: {this.props.event.date}</p>
-          <p>Created by {this.props.event.user.username}</p>
-          <EditModalHelper {...this.props} />
-          {/* <button type="button"
-            onClick={ () => { this.startModal() }}>Edit</button> */}
-          <button type="button" onClick={() => this.props.deleteEvent(this.props.event.id)}>Delete</button>
-        </div>
-      </div>
-    );
-  }
+    render() {
+        const currentUser = JSON.parse(sessionStorage.getItem("credentials"))
+        if (this.props.event.user.username === currentUser.username) {
+            return (
+                <div className="card">
+                    <div className="card-content">
+                        <h3>{this.props.event.eventName}</h3>
+                        <p>Location: {this.props.event.location}</p>
+                        <p>Date: {this.props.event.date}</p>
+                        <p>Created by {this.props.event.user.username}</p>
+                        <EditModalHelper {...this.props} />
+                        <button type="button" onClick={() => this.props.deleteEvent(this.props.event.id)}>Delete</button>
+                    </div>
+                </div>
+            );
+        } else {
+            return (
+                <div className="card">
+                    <div className="card-content">
+                        <h3>{this.props.event.eventName}</h3>
+                        <p>Location: {this.props.event.location}</p>
+                        <p>Date: {this.props.event.date}</p>
+                        <p>Created by {this.props.event.user.username}</p>
+                    </div>
+                </div>
+            );
+
+        }
+    }
 }
 
 export default EventCard;
